fix another character type

+++ AssassinAttack.h
@@ -10,11 +10,11 @@ class AssassinAttack : public AttackStrategy
 {
 public:
     AssassinAttack() : AttackStrategy(){};
-    int attackEnemy(Character *currCharacter, Enemy *currEnemy)
+    double attackEnemy(Character *currCharacter, Enemy *currEnemy)
     {
         return abs(((((((2 * currCharacter->getLevel()) / 5) + 2) * 35 * currCharacter->getAttack()) * currCharacter->getMaxHP() / currCharacter->getCurrentHP()) / currCharacter->getDefense() * 5) + 2);
     }
-    int defenseEnemy(Character *currCharacter, Enemy *currEnemy)
+    double defenseEnemy(Character *currCharacter, Enemy *currEnemy)
     {
         return abs(((((((2 * currCharacter->getLevel()) / 5) + 2) * 35 * currCharacter->getDefense()) * currCharacter->getMaxHP() / currCharacter->getCurrentHP()) / currEnemy->get_enemy_atk() * 5) + 2);
     }
